1) server package fix. API work again. 2)Now, in addition to our Mux cors is used. 3) Game service now can work without chat service(use an additional slice to save unsent messages). 4) Table Message now have field not_saved_id

+++ chat/proto/service.go
@@ -1,21 +1,37 @@
 package chat
 
 import (
+	re "github.com/go-park-mail-ru/2019_1_Escapade/internal/return_errors"
+	"github.com/go-park-mail-ru/2019_1_Escapade/internal/utils"
+
 	"context"
 
 	"database/sql"
 
 	//
 	_ "github.com/lib/pq"
-
-	"github.com/go-park-mail-ru/2019_1_Escapade/internal/utils"
 )
 
+// Service of chat
 type Service struct {
 	DB *sql.DB
 }
 
+//NewService Create new instance of service
+func NewService(db *sql.DB) *Service {
+	return &Service{
+		DB: db,
+	}
+}
+
+// CreateChat create chat with or without users.
+// Specify the type of chat and id received from the corresponding database table
+// Return id for this chat, save it. It must be transferred to any chat operations
 func (service *Service) CreateChat(ctx context.Context, chat *ChatWithUsers) (*ChatID, error) {
+	if chat == nil {
+		return &ChatID{}, re.InvalidMessage()
+	}
+
 	var (
 		tx     *sql.Tx
 		err    error
@@ -27,7 +43,7 @@ func (service *Service) CreateChat(ctx context.Context, chat *ChatWithUsers) (*C
 	}
 	defer tx.Rollback()
 
-	if ChatID, err = service.insertChat(tx, chat); err != nil {
+	if ChatID, err = service.insertChat(tx, chat.Type, chat.TypeId); err != nil {
 		return ChatID, err
 	}
 
@@ -42,37 +58,99 @@ func (service *Service) CreateChat(ctx context.Context, chat *ChatWithUsers) (*C
 	return ChatID, err
 }
 
+// GetChat get the ID of the chat, based on its type and the passed ID of this type
 func (service *Service) GetChat(ctx context.Context, chat *Chat) (*ChatID, error) {
+	if chat == nil {
+		return &ChatID{}, re.InvalidChatID()
+	}
+
 	id, err := service.getChat(chat)
 	if err == nil {
 		utils.Debug(false, "GetChat success:", chat.Type, chat.TypeId, id)
+	} else {
+		var (
+			tx *sql.Tx
+		)
+		if tx, err = service.DB.Begin(); err != nil {
+			return id, err
+		}
+		defer tx.Rollback()
+		if id, err = service.insertChat(tx, chat.Type, chat.TypeId); err != nil {
+			return id, err
+		}
+		if err = tx.Commit(); err != nil {
+			return id, err
+		}
 	}
+
 	return id, err
 }
 
+// AppendMessage append message to database
+// to work correctly, specify the ID of the chat(in the message) in which
+// the operation occurs
+// Return id for this message, save it. It must be transferred to any message
+// operations
 func (service *Service) AppendMessage(ctx context.Context, message *Message) (*MessageID, error) {
-	utils.Debug(false, "Message time", message.Time)
+	if message == nil {
+		return &MessageID{}, re.InvalidMessage()
+	}
+
+	if message.ChatId <= 0 {
+		return &MessageID{}, re.InvalidMessageID()
+	}
+
 	id, err := service.insertMessage(message)
 	if err == nil {
 		utils.Debug(false, "AppendMessage success")
 	}
 	return id, err
 }
+
+// AppendMessages append messages to database
 func (service *Service) AppendMessages(ctx context.Context, messages *Messages) (*MessagesID, error) {
+	if messages == nil {
+		return &MessagesID{}, re.InvalidMessage()
+	}
+
 	ids, err := service.insertMessages(messages)
 	if err == nil {
 		utils.Debug(false, "AppendMessages success")
 	}
 	return ids, err
 }
+
+// UpdateMessage update message in database
+// to work correctly, specify the ID of the message in which
+// the operation occurs
 func (service *Service) UpdateMessage(ctx context.Context, message *Message) (*Result, error) {
+	if message == nil {
+		return &Result{}, re.InvalidMessage()
+	}
+
+	if message.Id <= 0 {
+		return &Result{}, re.InvalidMessageID()
+	}
+
 	res, err := service.updateMessage(message)
 	if err == nil {
 		utils.Debug(false, "UpdateMessage success")
 	}
 	return res, err
 }
+
+// DeleteMessage delete message from database
+// to work correctly, specify the ID of the message in which
+// the operation occurs
 func (service *Service) DeleteMessage(ctx context.Context, message *Message) (*Result, error) {
+	if message == nil {
+		return &Result{}, re.InvalidMessage()
+	}
+
+	if message.Id <= 0 {
+		return &Result{}, re.InvalidMessageID()
+	}
+
 	res, err := service.deleteMessage(message)
 	if err == nil {
 		utils.Debug(false, "deleteMessage success")
@@ -80,7 +158,21 @@ func (service *Service) DeleteMessage(ctx context.Context, message *Message) (*R
 	return res, err
 }
 
+// InviteToChat invite user to the chat
+// to work correctly, specify user and id of the chat
 func (service *Service) InviteToChat(ctx context.Context, userInChat *UserInGroup) (*Result, error) {
+	if userInChat == nil {
+		return &Result{}, re.InvalidUser()
+	}
+
+	if userInChat.User == nil || userInChat.User.Id <= 0 {
+		return &Result{}, re.InvalidUser()
+	}
+
+	if userInChat.Chat == nil || userInChat.Chat.Id <= 0 {
+		return &Result{}, re.InvalidChatID()
+	}
+
 	var (
 		tx  *sql.Tx
 		err error
@@ -102,7 +194,22 @@ func (service *Service) InviteToChat(ctx context.Context, userInChat *UserInGrou
 
 	return &Result{Done: true}, err
 }
+
+// LeaveChat leave user from the chat
+// to work correctly, specify user and id of the chat
 func (service *Service) LeaveChat(ctx context.Context, userInChat *UserInGroup) (*Result, error) {
+	if userInChat == nil {
+		return &Result{}, re.InvalidUser()
+	}
+
+	if userInChat.User == nil || userInChat.User.Id <= 0 {
+		return &Result{}, re.InvalidUser()
+	}
+
+	if userInChat.Chat == nil || userInChat.Chat.Id <= 0 {
+		return &Result{}, re.InvalidChatID()
+	}
+
 	res, err := service.deleteUserInChat(userInChat)
 	if err == nil {
 		utils.Debug(false, "LeaveChat success")
@@ -110,7 +217,12 @@ func (service *Service) LeaveChat(ctx context.Context, userInChat *UserInGroup) 
 	return res, err
 }
 
+// GetChatMessages get all messages from the chad with specified id
 func (service Service) GetChatMessages(ctx context.Context, chatID *ChatID) (*Messages, error) {
+	if chatID == nil {
+		return &Messages{}, re.InvalidMessage()
+	}
+
 	messages, err := service.getChatMessages(chatID)
 	if err == nil {
 		utils.Debug(false, "GetChatMessages success. Id was", chatID.Value,
